Merge remote-tracking branch 'paparazzi/master' into messages

Conflicts:
	Makefile
	Makefile.install
	conf/messages.xml
	conf/telemetry/default_rotorcraft.xml
	conf/units.xml
	sw/airborne/booz/test/booz2_test_micromag.c
	sw/airborne/firmwares/vor/lpc_vor_main.c
	sw/airborne/lisa/test/lisa_test_ms2100.c
	sw/airborne/modules/sensors/imu_ppzuav.c
	sw/ground_segment/cockpit/compass.ml
	sw/ground_segment/cockpit/live.ml
	sw/ground_segment/joystick/input2ivy.ml
	sw/ground_segment/python/messages_app/messagesframe.py
	sw/ground_segment/tmtc/ivy2udp.ml
	sw/ground_segment/tmtc/ivy_tcp_aircraft.ml
	sw/ground_segment/tmtc/ivy_tcp_controller.ml
	sw/ground_segment/tmtc/link.ml
	sw/ground_segment/tmtc/messages.ml
	sw/ground_segment/tmtc/rotorcraft_server.ml
	sw/ground_segment/tmtc/server.ml
	sw/ground_segment/tmtc/server_globals.ml
	sw/ground_segment/tmtc/server_msg_v1.ml
	sw/lib/ocaml/papget.ml
	sw/lib/ocaml/pprz.ml
	sw/lib/python/messages_tool.py
	sw/lib/python/messages_xml_map.py
	sw/tools/gen_messages.ml

+++ sw/airborne/modules/nav/nav_catapult.h
@@ -22,7 +22,7 @@
  */
 
 /**
- * @file module/nav/nav_catapult.h
+ * @file modules/nav/nav_catapult.h
  * @brief catapult launch timing system
  */
 
@@ -32,9 +32,9 @@
 #include "std.h"
 #include "paparazzi.h"
 
-extern int nav_catapult_motor_delay;
+extern float nav_catapult_motor_delay;
 extern float nav_catapult_acceleration_threshold;
-extern int nav_catapult_heading_delay;
+extern float nav_catapult_heading_delay;
 extern float nav_catapult_initial_pitch;
 extern float nav_catapult_initial_throttle;
